+++ src/test/java/ladder/view/GameResultTests.java
@@ -1,5 +1,8 @@
-package ladder.domain;
+package ladder.view;
 
+import ladder.domain.Participant;
+import ladder.domain.Results;
+import ladder.view.GameResult;
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
